+++ src/sagemaker/local/local_session.py
@@ -17,7 +17,6 @@ import logging
 import platform
 
 import boto3
-import urllib3
 from botocore.exceptions import ClientError
 
 from sagemaker.local.image import _SageMakerContainer
@@ -29,7 +28,14 @@ from sagemaker.local.entities import (
     _LocalTransformJob,
 )
 from sagemaker.session import Session
-from sagemaker.utils import get_config_value
+from sagemaker.utils import DeferredError, get_config_value
+
+try:
+    import urllib3
+except ImportError as e:
+    logging.warning("urllib3 failed to import. Local mode features will be impaired or broken.")
+    # Any subsequent attempt to use urllib3 will raise the ImportError
+    urllib3 = DeferredError(e)
 
 logger = logging.getLogger(__name__)